+++ crates/impl-more/src/lib.rs
@@ -22,6 +22,7 @@
 //!
 //! impl_more::impl_from!(String => MyNewTypeStruct);
 //! impl_more::impl_into!(MyNewTypeStruct => String);
+//! impl_more::forward_from_str!(MyNewTypeStruct => String);
 //!
 //! enum MyEnum {
 //!     Bar,
@@ -100,6 +101,8 @@ mod display;
 mod error;
 #[macro_use]
 mod from;
+#[macro_use]
+mod from_str;
 
 #[cfg(test)]
 mod tests {