+++ ensure-no-std/src/compat_test.rs
@@ -13,6 +13,7 @@ impl_more::impl_deref_mut!(Foo);
 
 impl_more::impl_from!(String => Foo);
 impl_more::impl_into!(Foo => String);
+impl_more::forward_from_str!(Foo => String);
 
 impl_more::forward_display!(Foo);
 
@@ -37,6 +38,7 @@ struct Baz<T> {
     inner: T,
 }
 
+impl_more::forward_from_str!(<T> in Baz<T> => inner: T);
 impl_more::forward_display!(<T> in Baz<T> => inner);
 
 #[derive(Debug, Clone)]