+++ crates/impl-more/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add `forward_from_str!` macro.
+
 ## 0.3.4
 
 - Add `forward_as_ref!`, `forward_as_mut!`, and `forward_as_ref_and_mut!` macros.